pageable

+++ src/main/java/com/week_project/service/UtenteService.java
@@ -1,10 +1,13 @@
 package com.week_project.service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import com.week_project.model.Edificio;
@@ -13,6 +16,8 @@ import com.week_project.model.Stato_Postazione;
 import com.week_project.model.Utente;
 import com.week_project.repository.PostazioneDaoRepository;
 import com.week_project.repository.UtenteDaoRepository;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UtenteService {
@@ -47,7 +52,13 @@ public class UtenteService {
 		return repo.findByCityInCitta(stato,citta);
 	}
 	public String removeUtenteByID(Long id) {
+		if(!repo.existsById(id)) {
+			throw new NoSuchElementException("non c'e l'utente");
+		}
 		repo.deleteById(id);
 		return "utente eliminato";
+	}
+	public Page<Utente> findAllPageable(Pageable pageable){
+		return (Page<Utente>) repo.findAll(pageable);
 	}
 }
